+++ lib/sepa_king.rb
@@ -12,6 +12,7 @@ require 'sepa_king/transaction'
 require 'sepa_king/transaction/direct_debit_transaction'
 require 'sepa_king/transaction/credit_transfer_transaction'
 require 'sepa_king/message'
+require 'sepa_king/message/shared/postal_address_builder'
 require 'sepa_king/message/direct_debit'
 require 'sepa_king/message/credit_transfer'
 require 'sepa_king/postal_address'
